Refactor vesu protocol view.
